updated ft_substr.c and libft.h

+++ new_libft/ft_substr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_substr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pnourish <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: pnourish  <marvin@42.fr>                   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/10 20:53:22 by pnourish          #+#    #+#             */
-/*   Updated: 2023/05/26 20:50:20 by pnourish         ###   ########.fr       */
+/*   Updated: 2023/09/15 21:54:50 by pnourish         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,9 +20,9 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 
 	if (!s)
 		return (NULL);
-	if ((int)start >= ft_strlen(s))
+	if ((size_t)start >= ft_strlen(s))
 		return (ft_strdup(""));
-	if ((int)len > ft_strlen(s))
+	if ((size_t)len > ft_strlen(s))
 		len = ft_strlen(s);
 	str = (char *) malloc(sizeof(*s) * (len + 1));
 	if (!str)
